+++ packages/core/settings.js
@@ -22,7 +22,7 @@ Kepler.settings = {
 			"de": "Deutsch",
 			"sv": "Svenska",
 			"es": "Español",
-			"fr": "Français",
+			"fr": "Français"
 		},
 		"templates": {
 			//core/ui templates configurations
@@ -33,7 +33,16 @@ Kepler.settings = {
 				"navSidebar_ui_connection": { order:10 }
 			},
 			"panelProfile": {
-				"panelProfile_ui_hist": { order:-10 }
+				//"panelProfile_ui_tabloc": { order:-10 },
+				"panelProfile_ui_hist": { order:-10 },
+				"panelProfile_ui_sets": { order:10 }
+			},
+			"panelUser": {
+				"panelUser_ui_btns": { order:-10 },
+				"panelUser_ui_tabloc": { order:-10 }
+			},
+			"popupPlace": {
+				"popupPlace_ui_checkins": {order:-10}
 			},
 			"footer": {
 				"footer_ui_builtwith": { order:10 }