+++ packages/ui/client/views/panels/settings/bio.js
@@ -1,19 +1,5 @@
 
-Template.panelSettings.helpers({
-	genders: function() {
-		var gender = Meteor.user() && (Meteor.user().gender || 'none');
-		return _.map({male:'male',female:'female',none:'none'}, function(val, k) {
-			return {
-				key: k,
-				val: val,
-				name: i18n('gender_'+k),
-				active: gender===val
-			};
-		});
-	}
-});
-
-Template.panelSettings.events({
+Template.panelSettings_ui_bio.events({
 
 	'keyup #username': _.debounce(function(e, tmpl) {
 		var input$ = $(e.target),
@@ -61,11 +47,6 @@ Template.panelSettings.events({
 			Users.update(Meteor.userId(), { $set: {'name': val } });
 		}
 	}, 300),
-
-/*	'keyup #city': _.debounce(function(e) {
-		var val = $(e.currentTarget).val();
-		Users.update(Meteor.userId(), { $set: {'city': val } });
-	}, 300),*/
 	
 	'change #gender input': _.debounce(function(e) {
 		e.preventDefault();
@@ -74,27 +55,19 @@ Template.panelSettings.events({
 		Users.update(Meteor.userId(), { $set: {'gender': val } });
 
 	}, 300)
-	/*
-	//TODO
-	'keyup #email': _.debounce(function(e) {
-		var val = $(e.target).val(),
-			oldval = $(e.target).data('value'),
-			feed$ = $(e.target).next('.form-control-feedback');
 
-		if(!K.Util.valid.email(val))
-			feed$.show();
-		else {
-			feed$.hide();
-			//TODO
-			// Users.update(Meteor.userId(), {
-			// 	$set: {
-			// 		emails: [{
-			// 			address: val,
-			// 			verified: val==oldval
-			// 		}]
-			// 	}
-			// });
-		}
-	}, 300),
-	*/	
+});
+
+Template.panelSettings_ui_bio.helpers({
+	genders: function() {
+		var gender = Meteor.user() && (Meteor.user().gender || 'none');
+		return _.map({male:'male',female:'female',none:'none'}, function(val, k) {
+			return {
+				key: k,
+				val: val,
+				name: i18n('gender_'+k),
+				active: gender===val
+			};
+		});
+	}
 });