+++ packages/ui/client/views/panels/settings/block.js
@@ -1,5 +1,5 @@
 
-Template.item_user_blocked.events({
+Template.item_user_block.events({
 	'click .user-btn-unblock': function(e, tmpl) {
 		K.Profile.userUnblock(this.id);
 	}