+++ packages/lib/package.js
@@ -1,4 +1,4 @@
-var version = '1.4.9';
+var version = '1.5.0';
 
 Package.describe({
   version: version,