+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+- render ui/core templates by settings.templates option
 - new settings K.settings.public.accounts.services to disable each oauth service
 
 ## v1.4.9