+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 - new settings K.settings.public.accounts.services to disable each oauth service
 - redirections base paths /users, /places
 - split panel settings templates in more configurable templates
+- convert user lang code in lang name
 
 ## v1.4.9
 - new settings bboxMaxResults